Refactor: Move the existing APIv1 module to an appropriately named file
